Correct capitalization of environment variable name.
When I initially deployed this, I used an uppercase variable name as
later versions of the Prometheus client library (all of which I am using
in various of my projects) switched to uppercase and deprecated the old
way of specifying this variable. Unfortunately the `django-prometheus`
library uses a lowercase variable to check whether the multiprocess mode
is in use, so we were collecting per-worker metrics the entire time,
which completely messed up our time series.