Bump modmail resource requirements
Over the past 8 days modmail has restarted 17 times due to hitting resource limits.

As of the time of writing this, it is sitting at a steady 427Mi of memory usage.

I believe this change is needed due to the increased size of our server, as well as larger data set of previous logs to go through when searching (causing increased memory usage)